Add SpaceCadet emulation & variable release delay
This is a major rewrite, involving two related design changes. The bigger one is
that if a `Qukey` is defined with a _primary_ `Key` value that is a modifier
(including layer shifts), it will be treated like a SpaceCadet key, rather than
a normal qukey. The "hold" and "tap" values are flipped; the _primary_ value
gets used on "hold", and the _alternate_ value gets used on "tap". Additionally,
the primary (modifier) value will always get used if there is any amount of
rollover to another key; a SpaceCadet key must be released before any other key
is toggled on in order to collapse into its alternate state.

Second (this would have been a separate commit if it wasn't so closely related),
the release delay is now dynamic, and depends on the length of time the released
key has been held; basically, the release delay gets longer if there's more
overlap between the qukey and the subsequent key. The configuration parameter
now sets the ratio (in units of 10%) required for the qukey to collapse into its
alternate state if it is released before the subsequent key. A value of 10
(100%) means that the subsequent key must be released first (it must have 100%
overlap with the qukey), whereas a value of 5 means that it must have at least
50% overlap -- if it's held long enough that less than half of its keypress was
spent during the time that the qukey was held, the qukey will be in its primary
state.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>