Rewrite for general event queue
This is a major rewrite of Qukeys. Instead of just queueing key press events, we
now queue releases as well. This greatly simplifies the hook functions, and
makes the code substantially clearer, but it does seem to result in a larger
program binary.

The major functional improvement is that it fixes the problem of rollover from a
non-qukey modifier to a qukey. In that case, the modifier should get applied to
the qukey, but as the release event for the modifier would get ignored by qukeys
until after the qukey's press state was decided, it wouldn't.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>