Fix EventQueue::remove(), add EventQueue::shift()
I had neglected to set `release_event_bits_` correctly. With that fix, the
compiled size get bigger by enough that it makes more sense to add a simpler
`shift()` function that only removes the event at the head of the queue, since
that's what we almost always want, anyway. I left `remove()` in there for now,
because it might turn out to be useful for something.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>