Remove code for releasing non-modifiers out of order
This was a tough decision, but the increase in code size and complexity didn't
seem worth it for a couple of very unlikely corner cases. In one, we could get a
tapped non-modifier held long enough to repeat because of rollover to a qukey,
but even then, it would only happen with an atypically long timeout setting.

In the other, the queue would fill up and prematurely release the qukey (in its
primary state) where it would otherwise have stayed shorter by releasing the
non-modifier that rolled over to the qukey press. In that case, the user would
still have to press six keys before releasing any of them, all before the qukey
times out. I'm willing to let that case go.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>