Add tap-hold composite event feature
This feature allows a user to get a hold for the primary value of a Qukey (or
the alternate value of a SpaceCadet qukey) by quickly following a tap with a
press & hold.

It works by delaying the release event when a qukey is tapped. If the qukey is
pressed again before the timeout, the delay is extended (by the same amount),
and if the qukey is not released before that timeout, the first release and the
second press events are dropped from the queue without sending any reports to
the host, so a tap-hold (quick press-release-press-hold) turns into a single
press & hold in the keyboard's output.

Closes #1

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>